service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic>=2.11.7,<3.0.0
pydantic-settings>=2.3.0,<3.0.0
deepeval==3.6.9
requests==2.31.0
python-multipart==0.0.6
python-dotenv>=1.1.1,<2.0.0